Move rabbitmq implementation to separate file
